FindMate_V2/FindMate-ML-Service-v2/FindMate-ML-Service/requirements.txt
# FindMate ML Service - Python Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
pymongo==4.6.1

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.26.2

# Utilities
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Optional: For advanced NLP (uncomment if needed)
# spacy==3.7.2
